+++ features/top_ads/top_ads_autoads/src/main/java/com/tokopedia/topads/auto/view/viewmodel/DailyBudgetViewModel.kt
@@ -45,7 +45,7 @@ class DailyBudgetViewModel @Inject constructor(
 
     fun getBudgetInfo(requestType: String, source: String, onSuccess: (ResponseBidInfo.Result) -> Unit) {
         launchCatchError(block = {
-            val dummyId: MutableList<Int> = mutableListOf()
+            val dummyId: MutableList<Long> = mutableListOf()
             val suggestionsDefault = ArrayList<DataSuggestions>()
             suggestionsDefault.add(DataSuggestions(PRODUCT, dummyId))
             bidInfoUseCase.setParams(suggestionsDefault, requestType, source)
